comments and fixed with new database
got all the work done
